Consolidate diagnostic logic, avoid duplicating types etc
