+++ src/__tests__/language/diagnostics/linter.test.ts
@@ -4,7 +4,7 @@ import {
   rangeFromMessage,
   getWarningMessage,
   getErrorMessage,
-} from "../../../language/diagnostics";
+} from "../../../language/diagnostics/linter";
 
 // all messages actually generated by stanc
 const jacobianWarning = `Warning in 'jacobian.stan', line 1, column 12: Variable name 'jacobian' will
