--- src/language/diagnostics/index.ts
@@ -1,2 +1,0 @@
-export { provideDiagnostics } from "./provider";
-export { rangeFromMessage, getWarningMessage, getErrorMessage } from "./linter";
