+++ src/types/common.ts
@@ -5,25 +5,4 @@ export interface Position {
   character: number;
 }
 
-type StancSuccess = {
-  errors: undefined;
-  result: string;
-  warnings?: string[];
-};
-
-type StancFailure = {
-  errors: string[];
-  result: undefined;
-  warnings?: string[];
-};
-
-export type StancReturn = StancSuccess | StancFailure;
-
-export type StancFunction = (
-  filename: string,
-  code: string,
-  options: string[],
-  includes?: Record<string, string>,
-) => StancReturn;
-
 export type FileSystemReader = (filename: Filename) => Promise<FileContent>;
