+++ src/types/index.ts
@@ -1,4 +1,3 @@
 // Barrel export for all types
 export * from './common';
 export * from './completion';
-export * from './diagnostics';
